Addition of a new endpoint in the articles controller (needs to get verified)

+++ back/src/main/java/com/openclassrooms/p6/payload/response/ArticleSummary.java
@@ -3,10 +3,10 @@ package com.openclassrooms.p6.payload.response;
 import java.time.LocalDateTime;
 
 public record ArticleSummary(
-        int id,
-        int userId,
-        int articleId,
-        String title,
-        String description,
-        LocalDateTime publicationDate) {
+                Long id,
+                Long userId,
+                Long articleId,
+                String title,
+                String description,
+                LocalDateTime publicationDate) {
 }
